requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# blueapps requirement
Django==3.2.4
PyMySQL==0.6.7
mysqlclient==2.0.3
MarkupSafe==1.1.1
Mako==1.0.6
requests==2.22.0
celery==4.4.5
django-celery-beat==2.0.0
django-celery-results==2.0.0
# pi==3.3.1
python-json-logger==0.1.7
whitenoise==5.2.0
six==1.16.0
httplib2==0.9.1
djangorestframework==3.12.4
drf-extensions==0.7.0
django-cors-headers==3.7.0
jsonfield==3.1.0
pypinyin==0.31.0
django_extensions==2.1.0
django-filter==2.4.0
# django-autofixture==0.12.1
django-revproxy==0.10.0
pyinstrument_cext==0.2.2
pyinstrument==3.0.3
humanize==0.5.1
xlwt==1.3.0
jsonschema==3.2.0
# django-smart-autoregister==0.0.5
django-bulk-update==2.2.0
django-redis==5.0.0
RestrictedPython==5.0

# wiki related
bleach==1.5.0
django-classy-tags==2.0.0
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-nyt==1.2
django-sekizai==2.0.0
Markdown==3.2.1
Pillow==5.2.0
sorl-thumbnail==12.7.0
django-simplemde==0.1.2
martor==1.2.8
python-dateutil==2.7.5
django-multiselectfield==0.1.12

# monitor
statsd==3.3.0

# pipeline
ujson==1.35
pyparsing==2.2.0
redis==3.5.3
redis-py-cluster==2.1.3
django-timezone-field==4.1.2
factory_boy==2.11.1
mistune==0.8.4
eventlet==0.25.1
gevent==1.2.2

# iam
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
curlify==2.2.1
idna==2.8
urllib3==1.25.9
# pycrypto==2.6.1

#others
raven==6.1.0
ddtrace==0.14.1
gunicorn==19.6.0
bkstorages==1.0.1
pytz==2019.3
suds-jurko==0.6
mock==3.0.5

# excel
xlrd==1.2.0
typing-extensions==3.7.4.3

anyjson==0.3.3

#bkouth
PyJWT==1.6.4
cryptography==3.3.2

# prometheus
django-prometheus==2.1.0

# grpcio
grpcio==1.27.2

protobuf==3.19.4

# opentelemetry
opentelemetry-api==1.6.2
opentelemetry-sdk==1.6.2
opentelemetry-exporter-otlp==1.6.2
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-exporter-jaeger-proto-grpc==1.6.2
opentelemetry-exporter-jaeger-thrift==1.6.2
opentelemetry-instrumentation==0.25b2
opentelemetry-instrumentation-celery==0.25b2
opentelemetry-instrumentation-django==0.25b2
opentelemetry-instrumentation-dbapi==0.25b2
opentelemetry-instrumentation-redis==0.25b2
opentelemetry-instrumentation-logging==0.25b2
opentelemetry-instrumentation-requests==0.25b2

pyCryptodome==3.14.1

Jinja2==2.11.1
jmespath==0.10.0
requests_toolbelt==0.9.1

apigw-manager[cryptography]==1.0.7

drf-yasg==1.20.0
blueapps==4.2.3
django-bulk-update-or-create==0.3.0

# 打包上线依赖
supervisor==4.2.2
uwsgi==2.0.18
